rename snippet_helper_functions.ts -> parse_snippets.ts
